move config misc from utils to core
