+++ deltann/core/data.h
@@ -24,7 +24,7 @@ limitations under the License.
 #include "core/buffer.h"
 #include "core/io.h"
 #include "core/shape.h"
-#include "core/utils/misc.h"
+#include "core/misc.h"
 
 #ifdef USE_TF
 #include "tensorflow/core/framework/tensor_shape.h"
