+++ deltann/core/tfmodel.h
@@ -24,8 +24,8 @@ limitations under the License.
 
 #include "core/base_model.h"
 #include "core/data.h"
-#include "core/utils/logging.h"
-#include "core/utils/misc.h"
+#include "core/logging.h"
+#include "core/misc.h"
 
 #include "tensorflow/cc/saved_model/loader.h"
 #include "tensorflow/cc/saved_model/tag_constants.h"
